Reworked GpadButtonPropertyDrawer according to feedback.
Removed general AliasedEnumPropertyDrawer script.
